Update README.md
Updated the **Changelog** section (that just listed one version and was out of synch with the docs/CHANGELOG.md file) to include a link to the CHANGELOG.md file instead. Also updated the **Release** section to also refer to the CHANGELOG.md file instead of the existing link to bookmark (that was a 404).