fix: don't run on_serve if on_config was skipped
## Architecture
Since this plugin is not functional without nav, on_config detects
whether this needs to be run as a first step, and returns immediately if
it is not found.

However, currently on_serve fails with a TypeError if on_config returns
immediately, since it tries to server.watch(None). Therefore, we should
also skip on_serve if on_config never ran in the first place.

## Other changes
- Added tests folder inside of the plugin -- run with `python3 -m
  unittest` on the root of the repository